removed duplicated interfaces
