+++ types/events/events.types.ts
@@ -4,6 +4,12 @@ export enum MatchEventType {
   FailedPenalty = 40,
   YellowCard = 43,
   RedCard = 45,
+  // TODO COMPROBAR ESTO SI ES CIERTO XD
+  // Sustitución
+  // jugador entrante
+  // PlayerIn = 5,
+  // jugador saliente
+  // PlayerOut = 4,
   HalfTime,
   FinalTime,
   Overtime,
@@ -22,14 +28,7 @@ export const eventTypeLabels: Record<MatchEventType, string> = {
   [MatchEventType.None]: '',
 };
 
-export type ParsedMinute = {
-  min: number;
-  extra: number;
-  total: number;
-  label: string;
-};
-
-export type MatchEvent = {
+export interface MatchEvent {
   type: MatchEventType;
   minute?: number | string;
   extraMinute?: number;
@@ -37,17 +36,21 @@ export type MatchEvent = {
   playerName?: string;
   score?: string;
   extra?: string;
-};
+}
 
 export interface MatchData {
   id: number;
-  status: string; // TODO:  NS FT Canc. y mas
+  status: MatchStatus;
   result: string;
-  kickoff: Date;
+  kickoff: string; // 21:00 01/06/2025
   events?: MatchEvent[];
   homeId: number;
   awayId: number;
   competitionid: number;
+  homeTeam: TeamInfo;
+  awayTeam: TeamInfo;
+  country: string;
+  Odds: Odds;
 }
 
 export type TeamInfo = {
@@ -62,6 +65,42 @@ export interface MatchFullData extends MatchData {
   homeTeam: TeamInfo;
   awayTeam: TeamInfo;
 }
+
+export interface Odds {
+  id: string;
+  matchId: number;
+  homeOdd: number;
+  awayOdd: number;
+  drawOdd: number;
+}
+
+export const MATCH_STATUSES = [
+  'NS', // Not Started
+  'HT', // Half Time
+  'FT', // Full Time
+  'OT', // Overtime no se si existe
+  'AET', // After Extra Time
+  'AP', // After Penalties
+  'Canc.', // Cancelled
+] as const;
+
+export const FINAL_STATUSES: MatchStatus[] = ['FT', 'AET', 'AP', 'Canc.'];
+export type MatchStatus = (typeof MATCH_STATUSES)[number];
+
+export interface CompetitionData {
+  id: string;
+  name: string;
+  fullName: string;
+  badge: string;
+  matches: MatchData[];
+}
+
+export type ParsedMinute = {
+  min: number;
+  extra: number;
+  total: number;
+  label: string;
+};
 
 export interface EventsSportsResponse {
   football: {
